+++ WooCommerce/Classes/ViewRelated/Products/Edit Product/List Selector Data Source/PaginatedProductShippingClassListSelectorDataSource.swift
@@ -10,7 +10,7 @@ struct PaginatedProductShippingClassListSelectorDataSource: PaginatedListSelecto
     private let siteID: Int64
 
     init(product: Product, selected: ProductShippingClass?) {
-        self.siteID = Int64(product.siteID)
+        self.siteID = product.siteID
         self.selected = selected
     }
